backend/requirements.txt
fastapi==0.115.8
uvicorn==0.30.6
pydantic-settings==2.7.1
httpx==0.27.2
PyJWT[crypto]==2.9.0
google-auth==2.36.0
requests==2.32.3
